requirements.txt
numpy==1.21.0
sklearn~=0.0
scikit-learn~=1.0.2
tld~=0.12.6
beautifulsoup4~=4.10.0
liac-arff==2.5.0